% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{C2kivu.RData}
\alias{C2kivu.RData}
\title{CryoSat-2 water level observations for lake Kivu}
\description{
The CryoSat-2 L1b waveforms were retracked by the Narrow Primary Peak retracker (Villadsen, H., Andersen, O.B., Stenseng, L., Nielsen, K. and Knudsen, P., 2015. CryoSat-2 altimetry for river level monitoring—Evaluation in the Ganges–Brahmaputra River basin. Remote Sensing of Environment, 168, pp.80-89.). CryoSat-2 is available from
\itemize{
\item L1b SAR Precise Orbit. Baseline E.\url{https://doi.org/10.5270/CR2-fbae3cd}
\item L1b LRM Precise Orbit, Baseline E.\url{https://doi.org/10.5270/CR2-41ad749}
\item L1b SARin Precise Orbit. Baseline E.\url{https://doi.org/10.5270/CR2-6afef01}
}
}
\references{
Paper in prep
}
\author{
Karina Nielsen (karni@space.dtu.dk)
}
\keyword{data}
