% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeoidCorr.R
\name{getGeoidCorr}
\alias{getGeoidCorr}
\title{Predict the geoid correction error}
\usage{
getGeoidCorr(fit, newdat, UTM = TRUE)
}
\arguments{
\item{fit}{Object returned by getLakeGeoid()}

\item{newdat}{a data set at least including the columns named "lon" and "lat" with the coordinates where the corretion should be predicted.}

\item{UTM}{logic variable to indicate if the estimated field is in a UTM projection (FALSE means the field is in degrees)}
}
\description{
Predict the geoid correction error
}
\keyword{predict}
