% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kivu.RData}
\alias{kivu.RData}
\title{Level 2 Water Mask Raster Image 250m Data Product, Version C extracted for lake Kivu}
\description{
Selected columns of the Level 2 Water Mask Raster Image 250m Data Product, Version C. The data has been filtered by "darkwater_frac", large outliers were removed, and the data has been down sampled to reduce the volume.
}
\references{
Surface Water Ocean Topography (SWOT). 2024. SWOT Level 2 Water Mask Raster Image Data Product, Version C. Ver. C. PO.DAAC, CA, USA. Dataset accessed \link{2025-04-01} at \url{https://doi.org/10.5067/SWOT-RASTER-2.0}
}
\author{
SWOT Level 2 Water Mask Raster Image 250m Data Product, Version C (10.5067/SWOT-RASTER-2.0)
}
\keyword{data}
