% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{myshape.RData}
\alias{myshape.RData}
\title{Shapefile of the African lake Kivu}
\description{
A shapefile of the class "sf" for the African lake Kivu. The shapefile is taken from the SWOT Prior Lake Database.
}
\references{
Wang, J., Pottier, C., Cazals, C., Battude, M., Sheng, Y., Song, C., Sikder, M.S., Yang, X., Ke, L., Gosset, M. and Oliveira, R.R.A., 2023. The Surface Water and Ocean Topography Mission (SWOT) Prior Lake Database (PLD): Lake mask and operational auxiliaries. Authorea Preprints.\url{https://www.authorea.com/doi/full/10.22541/au.170258987.72387777}
}
\author{
Wang et al.
}
\keyword{data}
