% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeoidCorr.R
\name{plot.LakeGeoid}
\alias{plot.LakeGeoid}
\title{Plot an object returned by the function getLakeGeoid()}
\usage{
\method{plot}{LakeGeoid}(fit, zlim = NULL, dat = NULL, doSave = FALSE)
}
\arguments{
\item{fit}{Object returned by getLakeGeoid()}

\item{zlim}{vector with the zlim interval c(low,high), defined by the field "omega" if not specified}

\item{dat}{The raw water level data can be added to the plot}

\item{doSave}{plot is saved to a pdf file}
}
\description{
Plot an object returned by the function getLakeGeoid()
}
\examples{
plot(fit)
}
\keyword{plot}
