% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeoidCorr.R
\name{print.LakeGeoid}
\alias{print.LakeGeoid}
\title{Simple print  of convergence statutus}
\usage{
\method{print}{LakeGeoid}(x)
}
\arguments{
\item{x}{An object of class "LakeGeoid"}
}
\value{
Print the objective function and state convergence
}
\description{
This function presents a summary of the output
}
