% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeoidCorr.R
\name{summary.LakeGeoid}
\alias{summary.LakeGeoid}
\title{Summary of output}
\usage{
\method{summary}{LakeGeoid}(x)
}
\arguments{
\item{x}{An object of class "LakeGeoid"}
}
\value{
Summary of output
}
\description{
This function presents a summary of the output
}
